+++ service/src/integrationTest/java/uk/gov/hmcts/reform/fpl/controllers/OrdersNeededControllerTest.java
@@ -4,6 +4,10 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.OverrideAutoConfiguration;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
+import uk.gov.hmcts.reform.fpl.enums.OrderType;
+import uk.gov.hmcts.reform.fpl.model.CaseData;
+import uk.gov.hmcts.reform.fpl.model.GroundsForRefuseContactWithChild;
+import uk.gov.hmcts.reform.fpl.model.Orders;
 
 import java.util.List;
 import java.util.Map;
@@ -69,6 +73,26 @@ class OrdersNeededControllerTest extends AbstractCallbackTest {
         assertThat(response.getData().get("secureAccommodationOrderType")).isEqualTo("NO");
         assertThat(((Map<String, Object>) response.getData().get("orders")).get("secureAccommodationOrderSection"))
             .isEqualTo(null);
+    }
+
+    @Test
+    void shouldShowRefuseContactFieldWhenRefuseContactIsSelected() {
+        AboutToStartOrSubmitCallbackResponse response = postAboutToSubmitEvent(CaseData.builder()
+            .orders(Orders.builder().orderType(List.of(OrderType.REFUSE_CONTACT_WITH_CHILD)).build()).build());
+
+        assertThat(response.getData().get("refuseContactWithChildOrderType")).isEqualTo("YES");
+    }
+
+    @Test
+    @SuppressWarnings("unchecked")
+    void shouldRemoveRefuseContactDataWhenRefuseContactIsUnselected() {
+        AboutToStartOrSubmitCallbackResponse response = postAboutToSubmitEvent(CaseData.builder()
+            .orders(Orders.builder().orderType(List.of(OrderType.CARE_ORDER)).build())
+            .groundsForRefuseContactWithChild(GroundsForRefuseContactWithChild.builder()
+                .personHasContactAndCurrentArrangement("test").build()).build());
+
+        assertThat(response.getData().get("refuseContactWithChildOrderType")).isNull();
+        assertThat((Map<String, Object>) response.getData().get("groundsForRefuseContactWithChild")).isNullOrEmpty();
     }
 
     @Test
